+++ openqaoa/openqaoa-qiskit/backends/__init__.py
@@ -1,0 +1,3 @@
+from .devices import DeviceQiskit
+from .qaoa_qiskit_qpu import QAOAQiskitQPUBackend
+from .qaoa_qiskit_sim import (QAOAQiskitBackendShotBasedSimulator, QAOAQiskitBackendStatevecSimulator)
